--- lab1/__init__.py
@@ -1,1 +1,0 @@
-from .util import *
