+++ mitdeeplearning/__init__.py
@@ -1,1 +1,2 @@
 import mitdeeplearning.lab1
+import mitdeeplearning.util
